refactoring: queries
