+++ src/routes/graphql/schema.ts
@@ -1,6 +1,13 @@
 import { GraphQLSchema } from 'graphql';
-import { Queries } from './types/queries.js';
+import { Queries } from './queries.js';
+import { Mutations } from './mutations.js';
+import { PrismaClient } from '@prisma/client';
+
+export interface Context {
+  prisma: PrismaClient;
+}
 
 export const schema = new GraphQLSchema({
   query: Queries,
+  mutation: Mutations,
 });
