+++ src/routes/graphql/types/memberType.ts
@@ -3,16 +3,16 @@ import { GraphQLObjectType, GraphQLNonNull, GraphQLInt, GraphQLEnumType, GraphQL
 export const MemberType = new GraphQLObjectType({
   name: 'MemberType',
   fields: () => ({
-    id: { type: new GraphQLNonNull(IdType) },
+    id: { type: new GraphQLNonNull(MemberTypeId) },
     discount: { type: GraphQLFloat },
     postsLimitPerMonth: { type: GraphQLInt },
   }),
 });
 
-const IdType = new GraphQLEnumType({
-    name: 'MemberTypeId',
+export const MemberTypeId = new GraphQLEnumType({
+    name: 'memberTypeId',
     values: {
-        BASIC: { value: 'BASIC' },
-        BUSINESS: { value: 'BUSINESS' },
-    }
-})
+      BASIC: { value: 'BASIC' },
+      BUSINESS: { value: 'BUSINESS' },
+    },
+  });
