+++ src/routes/graphql/types/profileType.ts
@@ -1,13 +1,25 @@
-import { GraphQLObjectType, GraphQLNonNull, GraphQLString, GraphQLBoolean, GraphQLInt } from 'graphql';
+import {
+  GraphQLObjectType,
+  GraphQLNonNull,
+  GraphQLString,
+  GraphQLBoolean,
+  GraphQLInt,
+} from 'graphql';
+
 import { UUIDType } from './uuid.js';
+import { MemberType } from './memberType.js';
 
 export const ProfileType = new GraphQLObjectType({
   name: 'ProfileType',
   fields: () => ({
     id: { type: new GraphQLNonNull(UUIDType) },
     isMale: { type: GraphQLBoolean },
-    yearOfBirth: { type: GraphQLInt},
-    userId: { type: new GraphQLNonNull(UUIDType) },
+    yearOfBirth: { type: GraphQLInt },
     memberTypeId: { type: GraphQLString },
+    userId: { type: new GraphQLNonNull(UUIDType) },
+    memberType: {
+      type: MemberType,
+      resolve: (parent) => parent.memberType ?? null,
+    },
   }),
 });
